Worked on ability points system to make resource consumption work. Having issues with properties dictionary in Land objects.

+++ Assets/Scripts/Action.cs
@@ -12,7 +12,7 @@ public abstract class Action
     /// Stores which resources are spent, and the amount spent
     /// </summary>
     public Dictionary<string, int> resourceCosts = new Dictionary<string, int>();
-    public int timeCost;
+    public float timeCost;
 
     /// <summary>
     /// Performs the action.
