+++ Assets/Scripts/ResourceEditor.cs
@@ -18,7 +18,7 @@ public class ResourceEditor
         resource = _resource;
     }
 
-    public void setLevel(int resourceLevel)
+    public void setLevel(float resourceLevel)
     {
         resource.currentLevel = resourceLevel;
     }
@@ -27,7 +27,7 @@ public class ResourceEditor
     /// Set amount of health drained per healthUpdate when below deficiency threshold.
     /// </summary>
     /// <param name="drainAmt">Amount of health drained per healthUpdate when below deficiency threshold.</param>
-    public void setDeficiencyHealthDrain(int drainAmt)
+    public void setDeficiencyHealthDrain(float drainAmt)
     {
         resource.deficiencyHealthDrain = drainAmt;
     }
@@ -36,7 +36,7 @@ public class ResourceEditor
     /// Set threshold at which deficiency occurs (health drain takes effect).
     /// </summary>
     /// <param name="deficiencyThreshold">Threshold at which deficiency occurs (health drain takes effect).</param>
-    public void setDeficiencyThreshold(int deficiencyThreshold)
+    public void setDeficiencyThreshold(float deficiencyThreshold)
     {
         resource.deficiencyThreshold = deficiencyThreshold;
     }
@@ -45,7 +45,7 @@ public class ResourceEditor
     /// Set health gained when health gain threshold is met.
     /// </summary>
     /// <param name="healthGain">Health gained when health gain threshold is met.</param>
-    public void setHealthGain(int healthGain)
+    public void setHealthGain(float healthGain)
     {
         resource.healthGain = healthGain;
     }
@@ -54,17 +54,17 @@ public class ResourceEditor
     /// Set threshold at which health is gained when it's surpassed.
     /// </summary>
     /// <param name="gainThreshold">Threshold at which health is gained when it's surpassed.</param>
-    public void setHealthGainThreshold(int gainThreshold)
+    public void setHealthGainThreshold(float gainThreshold)
     {
         resource.healthGainThreshold = gainThreshold;
     }
 
-    public void setMaxLevel(int maxLevel)
+    public void setMaxLevel(float maxLevel)
     {
         resource.maxLevel = maxLevel;
     }
 
-    public void setBaseUsage(int baseUsage)
+    public void setBaseUsage(float baseUsage)
     {
         resource.baseUsage = baseUsage;
     }
